refactor unused cm model code
